+++ scripts/main.js
@@ -1,6 +1,6 @@
 'use strict';
 const storyController = require("./storyController");
-const act1scene1 = require("./act1scene1");
+const act1scene1 = require("../story/actOne/act1scene1");
 const charactersView = require("./charactersView");
 const sceneFactory = require('./sceneFactory');
