+++ story/actOne/act1scene3.js
@@ -1,9 +1,9 @@
 'use strict';
 
-const actTwo = require("./act2scene1");
-const endings = require("./endings");
-const characters = require("./characters");
-const characterController = require("./characterController");
+const actTwo = require("../actTwo/act2scene1");
+const endings = require("../endings/endings");
+const characters = require("../characters/characters");
+const characterController = require("../../scripts/characterController");
 
 let askRoderigoForMoney = {
     name: "askRoderigoForMoney",
